feat(portal): modularize bundle asset pipeline
