+++ config/portal_asset_manifest.json
@@ -4,6 +4,10 @@
       "repo_path": "public/portal-assets/portal.css",
       "media_type": "text/css; charset=utf-8"
     },
+    "shared-ui-tokens.css": {
+      "repo_path": "public/portal-assets/shared-ui-tokens.css",
+      "media_type": "text/css; charset=utf-8"
+    },
     "portal.js": {
       "repo_path": "public/portal-assets/portal.js",
       "media_type": "text/javascript; charset=utf-8"
