+++ public/portal-assets/portal.css
@@ -25,6 +25,8 @@
  * 1. FONTS
  * ============================================================================ */
 
+@import url("__PORTAL_SHARED_TOKENS_URL__");
+
 @font-face {
     font-family: "Portal Sans";
     src: url("__PORTAL_FONT_SANS_URL__") format("woff2");
@@ -57,9 +59,6 @@ body {
     --ux-focus-shadow: rgba(8, 145, 178, 0.18);
     --ux-panel-border: rgba(100, 116, 139, 0.3);
     --ux-panel-border-strong: rgba(8, 145, 178, 0.3);
-    --ux-target-min-size: 44px;
-    --ux-label-size: 0.8125rem;
-    --ux-body-size: 1rem;
     --shell-ink: #0f172a;
     --shell-muted: #475569;
     --shell-border: var(--ux-panel-border);
