+++ public/portal-assets/shared-ui-tokens.css
@@ -1,0 +1,6 @@
+/* Shared UI primitives used by both the managed frontdoor and portal console. */
+:root {
+  --ux-target-min-size: 44px;
+  --ux-label-size: 0.8125rem;
+  --ux-body-size: 1rem;
+}
