+++ web/secure-landing/package.json
@@ -5,11 +5,12 @@
   "type": "module",
   "scripts": {
     "preinstall": "node ./scripts/check-node-version.mjs",
-    "dev": "node ./scripts/guard-runtime.mjs && next dev",
-    "build": "node ./scripts/guard-runtime.mjs && next build",
+    "build:portal": "node ./scripts/guard-runtime.mjs && node ./scripts/build-portal-bundle.mjs",
+    "dev": "npm run build:portal && node ./scripts/guard-runtime.mjs && next dev",
+    "build": "npm run build:portal && node ./scripts/guard-runtime.mjs && next build",
     "start": "node ./scripts/guard-runtime.mjs && node ./scripts/start-standalone.mjs",
     "seed-frontdoor-user": "node ./scripts/guard-runtime.mjs && node ./scripts/seed-frontdoor-user.mjs",
-    "test": "node ./scripts/guard-runtime.mjs && node --experimental-specifier-resolution=node --test tests/*.test.mjs"
+    "test": "npm run build:portal && node ./scripts/guard-runtime.mjs && node --experimental-specifier-resolution=node --test tests/*.test.mjs"
   },
   "engines": {
     "node": ">=22 <23"
@@ -21,5 +22,8 @@
     "next": "16.2.3",
     "react": "^19.2.0",
     "react-dom": "^19.2.0"
+  },
+  "devDependencies": {
+    "esbuild": "^0.28.0"
   }
 }
