+++ web/secure-landing/portal-src/internal/bootstrap-auth.js
@@ -1,0 +1,19 @@
+export function defaultPortalBootstrapPayload() {
+  return {
+    authMode: "managed_unavailable",
+    csrfToken: null,
+    actor: null,
+    features: {
+      apiKeyInput: false,
+      directDebug: false
+    }
+  };
+}
+
+export function isManagedAuthMode(authState) {
+  return Boolean(authState && authState.mode !== "direct_debug");
+}
+
+export function isManagedUnavailableMode(authState) {
+  return Boolean(authState && authState.mode === "managed_unavailable");
+}
