+++ web/secure-landing/portal-src/portal.template.js
@@ -81,403 +81,7 @@ let healthCheckInFlight = false;
 let lastHealthCheckAt = 0;
 let configPreviewTimerId = null;
 
-var __PortalInternal = (() => {
-  var __defProp = Object.defineProperty;
-  var __getOwnPropDesc = Object.getOwnPropertyDescriptor;
-  var __getOwnPropNames = Object.getOwnPropertyNames;
-  var __hasOwnProp = Object.prototype.hasOwnProperty;
-  var __export = (target, all) => {
-    for (var name in all)
-      __defProp(target, name, { get: all[name], enumerable: true });
-  };
-  var __copyProps = (to, from, except, desc) => {
-    if (from && typeof from === "object" || typeof from === "function") {
-      for (let key of __getOwnPropNames(from))
-        if (!__hasOwnProp.call(to, key) && key !== except)
-          __defProp(to, key, { get: () => from[key], enumerable: !(desc = __getOwnPropDesc(from, key)) || desc.enumerable });
-    }
-    return to;
-  };
-  var __toCommonJS = (mod) => __copyProps(__defProp({}, "__esModule", { value: true }), mod);
-
-  // web/secure-landing/portal-src/internal/index.js
-  var index_exports = {};
-  __export(index_exports, {
-    buildManagedArtifactUrl: () => buildManagedArtifactUrl,
-    createDomContract: () => createDomContract,
-    createPortalAuthState: () => createPortalAuthState,
-    createPortalBootstrapState: () => createPortalBootstrapState,
-    createPortalConfigState: () => createPortalConfigState,
-    createPortalLastDiagnosticsState: () => createPortalLastDiagnosticsState,
-    createPortalMetadataState: () => createPortalMetadataState,
-    createPortalPreviewState: () => createPortalPreviewState,
-    createPortalReadinessState: () => createPortalReadinessState,
-    createPortalRouteHelpers: () => createPortalRouteHelpers,
-    createPortalUiState: () => createPortalUiState,
-    createRenderScheduler: () => createRenderScheduler,
-    createRenderSurfaceRegistry: () => createRenderSurfaceRegistry,
-    defaultPortalBootstrapPayload: () => defaultPortalBootstrapPayload,
-    finalizeManagedAssetUrl: () => finalizeManagedAssetUrl,
-    isManagedAuthMode: () => isManagedAuthMode,
-    isManagedUnavailableMode: () => isManagedUnavailableMode,
-    isNativeEventSourceHandle: () => isNativeEventSourceHandle,
-    nativeEventSourceReadyState: () => nativeEventSourceReadyState,
-    shouldEnableDomAssertions: () => shouldEnableDomAssertions
-  });
-
-  // web/secure-landing/portal-src/internal/artifact-security.js
-  function finalizeManagedAssetUrl(parsedUrl) {
-    return `${parsedUrl.pathname}${parsedUrl.search}`;
-  }
-  function buildManagedArtifactUrl({
-    job,
-    artifact,
-    apiBase,
-    artifactLabel,
-    sanitizeManagedAssetUrl
-  }) {
-    const directUrl = String(artifact?.url || "").trim();
-    if (directUrl) return sanitizeManagedAssetUrl(`${apiBase}${directUrl}`);
-    if (!job || !artifact) return "";
-    const relativePath = artifactLabel(artifact);
-    if (!relativePath) return "";
-    const encodedSegments = relativePath.split("/").map((segment) => encodeURIComponent(segment)).join("/");
-    return sanitizeManagedAssetUrl(
-      `${apiBase}/v1/jobs/${encodeURIComponent(String(job.id || ""))}/artifacts/${encodedSegments}`
-    );
-  }
-
-  // web/secure-landing/portal-src/internal/bootstrap-auth.js
-  function defaultPortalBootstrapPayload() {
-    return {
-      authMode: "managed_unavailable",
-      csrfToken: null,
-      actor: null,
-      features: {
-        apiKeyInput: false,
-        directDebug: false
-      }
-    };
-  }
-  function isManagedAuthMode(authState) {
-    return Boolean(authState && authState.mode !== "direct_debug");
-  }
-  function isManagedUnavailableMode(authState) {
-    return Boolean(authState && authState.mode === "managed_unavailable");
-  }
-
-  // web/secure-landing/portal-src/internal/dom-contract.js
-  function buildRequiredElementError(method, selector) {
-    return new Error(`Portal DOM contract missing required ${method}: ${selector}`);
-  }
-  function shouldEnableDomAssertions(windowRef) {
-    const host = String(windowRef?.location?.hostname || "").trim().toLowerCase();
-    return host === "localhost" || host === "127.0.0.1" || host.endsWith(".local");
-  }
-  function createDomContract(documentRef, { devAssertions = false } = {}) {
-    function id(elementId, { required = false } = {}) {
-      const node = documentRef.getElementById(String(elementId || ""));
-      if (!node && required && devAssertions) {
-        throw buildRequiredElementError("id", `#${elementId}`);
-      }
-      return node;
-    }
-    function query(selector, { required = false } = {}) {
-      const node = documentRef.querySelector(String(selector || ""));
-      if (!node && required && devAssertions) {
-        throw buildRequiredElementError("selector", selector);
-      }
-      return node;
-    }
-    function assertPresent(elements, requiredKeys) {
-      if (!devAssertions) return;
-      const missingKeys = [];
-      for (const key of requiredKeys) {
-        if (!elements?.[key]) {
-          missingKeys.push(String(key));
-        }
-      }
-      if (missingKeys.length) {
-        throw new Error(`Portal DOM contract missing required elements: ${missingKeys.join(", ")}`);
-      }
-    }
-    return {
-      id,
-      query,
-      assertPresent
-    };
-  }
-
-  // web/secure-landing/portal-src/internal/render-surfaces.js
-  function createRenderScheduler(windowRef) {
-    const scheduleFrame = typeof windowRef?.requestAnimationFrame === "function" ? windowRef.requestAnimationFrame.bind(windowRef) : (callback) => windowRef.setTimeout(callback, 0);
-    return {
-      schedule(callback) {
-        scheduleFrame(callback);
-      }
-    };
-  }
-  function createRenderSurfaceRegistry() {
-    const surfaces = /* @__PURE__ */ new Map();
-    function normalizeSurface(surface = {}) {
-      return {
-        init: typeof surface.init === "function" ? surface.init : () => {
-        },
-        render: typeof surface.render === "function" ? surface.render : () => {
-        }
-      };
-    }
-    return {
-      register(name, surface) {
-        if (!name) return;
-        surfaces.set(String(name), normalizeSurface(surface));
-      },
-      init(state) {
-        for (const surface of surfaces.values()) {
-          surface.init(state);
-        }
-      },
-      render(name, state) {
-        const surface = surfaces.get(String(name));
-        if (!surface) return;
-        surface.render(state);
-      }
-    };
-  }
-
-  // web/secure-landing/portal-src/internal/routing.js
-  function createPortalRouteHelpers(windowRef) {
-    function build({
-      viewName,
-      jobId = "",
-      artifactPath = "",
-      compareEnabled = null,
-      resolveView,
-      normalizeSelectedJobId,
-      normalizeArtifactRoutePath,
-      activeContext
-    }) {
-      const url = new URL(windowRef.location.href);
-      const resolvedView = resolveView(viewName);
-      const resolvedJobId = normalizeSelectedJobId(jobId);
-      url.searchParams.set("view", resolvedView);
-      if ((resolvedView === "operate" || resolvedView === "review") && resolvedJobId) {
-        const context = activeContext(resolvedJobId);
-        const resolvedArtifactPath = normalizeArtifactRoutePath(artifactPath) || context.artifactPath;
-        const resolvedCompareEnabled = compareEnabled === null ? context.compareEnabled : Boolean(compareEnabled);
-        url.searchParams.set("job", resolvedJobId);
-        if (resolvedArtifactPath) {
-          url.searchParams.set("artifact", resolvedArtifactPath);
-        } else {
-          url.searchParams.delete("artifact");
-        }
-        if (resolvedCompareEnabled) {
-          url.searchParams.set("compare", "1");
-        } else {
-          url.searchParams.delete("compare");
-        }
-      } else {
-        url.searchParams.delete("job");
-        url.searchParams.delete("artifact");
-        url.searchParams.delete("compare");
-      }
-      return url;
-    }
-    function read({
-      resolveView,
-      normalizeSelectedJobId,
-      normalizeArtifactRoutePath,
-      normalizeCompareQueryValue
-    }) {
-      const url = new URL(windowRef.location.href);
-      return {
-        view: resolveView(url.searchParams.get("view")),
-        jobId: normalizeSelectedJobId(url.searchParams.get("job")),
-        artifactPath: normalizeArtifactRoutePath(url.searchParams.get("artifact")),
-        compareEnabled: normalizeCompareQueryValue(url.searchParams.get("compare")),
-        hasArtifact: url.searchParams.has("artifact"),
-        hasCompare: url.searchParams.has("compare")
-      };
-    }
-    return {
-      build,
-      read
-    };
-  }
-
-  // web/secure-landing/portal-src/internal/state.js
-  function createPortalConfigState() {
-    return {
-      preset: "premium",
-      inputDir: "./input_images",
-      outputDir: "./output/lux_depth_v3_apex",
-      qualityTier: "premium",
-      depthBackend: "da3",
-      depthDevice: "cpu",
-      segmentation: {
-        enable: false,
-        backend: "stub",
-        sam2ModelSize: "base",
-        sam2CheckpointPath: "",
-        strict: false
-      },
-      flags: {
-        materials: true,
-        pbr: true,
-        cache: true,
-        overwrite: false,
-        enableV2: false,
-        saveFloatDepth: false,
-        forceDepth: false,
-        strictInputs: false,
-        verifyImages: false,
-        allowSemanticFallback: false,
-        verbose: false,
-        quiet: false
-      },
-      v2Preset: "default",
-      emits: {
-        master16: true,
-        upscaled16: true,
-        marketing: false,
-        report: true,
-        runCard: true,
-        runCardVersion: "v1",
-        runCardIncludeProofs: false
-      },
-      gate: { archiveIndex: "", manifestJsonl: "" },
-      licenses: { nonCommercialOk: false, acceptApple: false, acceptResearchTools: false },
-      reconstruction: {
-        enable: false,
-        groupingMode: "single",
-        camerasSidecarPath: "",
-        iterations: 1e3,
-        tier: "apex_research",
-        emitSceneDebugBundle: false
-      },
-      raw: {
-        ingestMode: "auto",
-        wbMode: "camera",
-        demosaic: "AHD"
-      },
-      runtime: {
-        maxWorkersMode: "auto",
-        maxWorkers: "",
-        maxGpuWorkersMode: "auto",
-        maxGpuWorkers: "",
-        logLevel: ""
-      }
-    };
-  }
-  function createPortalMetadataState() {
-    return {
-      pipeline: "",
-      fields: {},
-      estimate_bands: {},
-      debug_bundle_policy: {},
-      advanced_sections: [],
-      backend_catalog: {}
-    };
-  }
-  function createPortalPreviewState() {
-    return {
-      pipeline: "",
-      requestKey: "",
-      status: "idle",
-      field_errors: [],
-      field_warnings: [],
-      inactive_fields: [],
-      normalized_args: {},
-      execution_args: {},
-      submitted_args: {},
-      readiness: null,
-      estimate_summary: null,
-      debug_bundle_summary: null,
-      next_best_action: null,
-      argv_preview: "",
-      error: "",
-      error_reason: "",
-      error_status: 0
-    };
-  }
-  function createPortalUiState() {
-    return {
-      debugBundleAcknowledged: false,
-      effectiveConfigOpen: false,
-      debugBundleGuardrailSeen: false,
-      buildStep: 1,
-      lastOverlayTrigger: null,
-      lastSelectedJobId: "",
-      disclosurePrefs: {
-        advanced: null,
-        governance: null,
-        reconstruction: null,
-        dispatchTools: false
-      }
-    };
-  }
-  function createPortalReadinessState() {
-    return {
-      server: {},
-      pipelines: {}
-    };
-  }
-  function createPortalAuthState() {
-    return {
-      mode: "managed_unavailable",
-      csrfToken: "",
-      actor: null,
-      features: {
-        apiKeyInput: false,
-        directDebug: false
-      }
-    };
-  }
-  function createPortalBootstrapState(now = Date.now()) {
-    const lastTransitionAt = Number.isFinite(Number(now)) ? Number(now) : Date.now();
-    return {
-      status: "pending",
-      lastErrorReason: "",
-      lastHttpStatus: 0,
-      activeController: null,
-      activeTimeoutId: null,
-      lastTransitionAt,
-      lastHealthEndpointPath: "",
-      pendingOnlineFollowup: false,
-      onlineFollowupComplete: false,
-      retry: {
-        timer: null,
-        attempt: 0,
-        deadlineAt: 0,
-        lastDelayMs: 0,
-        lastReason: "",
-        lastHttpStatus: 0,
-        lastAttemptAt: 0,
-        lastOutcome: "",
-        lastEventAt: 0
-      }
-    };
-  }
-  function createPortalLastDiagnosticsState() {
-    return {
-      warnings: [],
-      expectedOutputs: [],
-      healthState: "good",
-      healthLabel: "good"
-    };
-  }
-
-  // web/secure-landing/portal-src/internal/stream-transport.js
-  function isNativeEventSourceHandle(handle) {
-    return Boolean(handle && typeof handle.readyState === "number" && typeof handle.addEventListener === "function");
-  }
-  function nativeEventSourceReadyState(handle) {
-    if (!isNativeEventSourceHandle(handle)) return null;
-    const readyState = Number(handle.readyState);
-    return Number.isInteger(readyState) ? readyState : null;
-  }
-  return __toCommonJS(index_exports);
-})();
+/* __PORTAL_INTERNALS__ */
 
 const portalInternals = __PortalInternal;
 const portalRoute = portalInternals.createPortalRouteHelpers(window);
