+++ web/secure-landing/public/frontdoor-homepage.css
@@ -1,11 +1,10 @@
+@import url("/shared-ui-tokens.css");
+
 body.frontdoor-homepage {
   --ux-focus-ring: rgba(103, 232, 249, 0.96);
   --ux-focus-shadow: rgba(34, 211, 238, 0.24);
   --ux-panel-border: rgba(148, 163, 184, 0.24);
   --ux-panel-border-strong: rgba(103, 232, 249, 0.28);
-  --ux-target-min-size: 44px;
-  --ux-label-size: 0.8125rem;
-  --ux-body-size: 1rem;
   --continuity-panel: rgba(12, 18, 34, 0.76);
   --continuity-panel-strong: rgba(7, 12, 26, 0.94);
   --continuity-panel-soft: rgba(15, 23, 42, 0.7);
